Factor out key bindings from overlay component
